refactor: 将 `YumeSignEnum` 移动到 `constants/` 目录 & 更新导入
- 将 `YumeSignEnum.ts` 从 `lib/xboyPackage` 目录移动到 `constants/` 目录
- 同步更新相关文件中的导入路径
- 计划将常量定义放在统一位置
