+++ src/plugins/task.ts
@@ -1,7 +1,7 @@
 //@ts-nocheck
 import type { SimulatedPlayer } from '@minecraft/server-gametest'
 import {simulatedPlayers, testWorldLocation} from './main'
-import SIGN from '../lib/xboyPackage/YumeSignEnum'
+import SIGN from '../constants/YumeSignEnum'
 import type { EntityHealthComponent, Vector3 } from '@minecraft/server'
 import { system, world } from '@minecraft/server'
 import { getEntitiesNear, getPlayerNear } from '../utils/Util'
